Just int and atoll

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -9,7 +9,7 @@
  */
 int _atoi(char *s)
 {
-	long long int num;
+	int num;
 
 	num = atoll(s);
 	return (num);
